requirements.txt
openpyxl
python-pptx
requests
edgartools
playwright
pypdf
